update
git-svn-id: https://svn.rueckgr.at/private/diplomarbeit@2066 95840bc9-f7e0-4b9f-abbd-54bb5795df2f

+++ WeatherReader/src/weatherreader/model/Weather.java
@@ -181,7 +181,7 @@ public class Weather {
 			else {
 				nextStartTime = 2*weatherReports.get(a).getStartTime().getTime() - weatherReports.get(a-1).getStartTime().getTime();
 			}
-			weatherReports.get(a).setEndTime(new Interval("interval" + nextStartTime, nextStartTime));
+			weatherReports.get(a).setEndTime(new Interval("interval" + (int)nextStartTime, nextStartTime));
 		}
 		
 		printWeatherReports("Weather reports after normalization", weatherReports);
